requirements.txt
python-binance
fastapi
uvicorn[standard]
python-dotenv
pandas
ta
numpy